refactor : web 도메인을 상위 계층으로 이동
